send fixtures to conftest
